# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sneaker Store API",
    "version": "1.0.0",
    "description": "API documentation for the Sneaker Store project"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "email": { "type": "string" },
                      "address": { "type": "string" },
                      "phone": { "type": "string" }
                    },
                    "required": ["_id", "name", "email", "address", "phone"]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" }
                },
                "required": ["name", "email", "address", "phone"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User  created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "address": { "type": "string" },
                    "phone": { "type": "string" }
                  },
                  "required": ["_id", "name", "email", "address", "phone"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to retrieve",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User  found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "address": { "type": "string" },
                    "phone": { "type": "string" }
                  },
                  "required": ["_id", "name", "email", "address", "phone"]
                }
              }
            }
          },
          "404": {
            "description": "User  not found"
          }
        }
      },
      "put": {
        "summary": "Update a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" }
                },
                "required": ["name", "email", "address", "phone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User  updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "address": { "type": "string" },
                    "phone": { "type": "string" }
                  },
                  "required": ["_id", "name", "email", "address", "phone"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "User  not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "User  deleted"
          },
          "404": {
            "description": "User  not found"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "summary": "Get all categories",
        "responses": {
          "200": {
            "description": "A list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "description": { "type": "string" }
                    },
                    "required": ["_id", "name", "description"]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                },
                "required": ["name", "description"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "description": { "type": "string" }
                  },
                  "required": ["_id", "name", "description"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "summary": "Get a category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the category to retrieve",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "description": { "type": "string" }
                  },
                  "required": ["_id", "name", "description"]
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "summary": "Update a category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the category to update",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                },
                "required": ["name", "description"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "description": { "type": "string" }
                  },
                  "required": ["_id", "name", "description"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the category to delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted"
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
    "/api/sneakers": {
      "get": {
        "summary": "Get all sneakers",
        "responses": {
          "200": {
            "description": "A list of sneakers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "brand": { "type": "string" },
                      "category_id": { "type": "integer" },
                      "price": { "type": "number" },
                      "stock": { "type": "integer" }
                    },
                    "required": ["_id", "name", "brand", "category_id", "price", "stock"]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new sneaker",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "brand": { "type": "string" },
                  "category_id": { "type": "integer" },
                  "price": { "type": "number" },
                  "stock": { "type": "integer" }
                },
                "required": ["name", "brand", "category_id", "price", "stock"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sneaker created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "brand": { "type": "string" },
                    "category_id": { "type": "integer" },
                    "price": { "type": "number" },
                    "stock": { "type": "integer" }
                  },
                  "required": ["_id", "name", "brand", "category_id", "price", "stock"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/sneakers/{id}": {
      "get": {
        "summary": "Get a sneaker by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the sneaker to retrieve",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Sneaker found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "brand": { "type": "string" },
                    "category_id": { "type": "integer" },
                    "price": { "type": "number" },
                    "stock": { "type": "integer" }
                  },
                  "required": ["_id", "name", "brand", "category_id", "price", "stock"]
                }
              }
            }
          },
          "404": {
            "description": "Sneaker not found"
          }
        }
      },
      "put": {
        "summary": "Update a sneaker by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the sneaker to update",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "brand": { "type": "string" },
                  "category_id": { "type": "integer" },
                  "price": { "type": "number" },
                  "stock": { "type": "integer" }
                },
                "required": ["name", "brand", "category_id", "price", "stock"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sneaker updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "brand": { "type": "string" },
                    "category_id": { "type": "integer" },
                    "price": { "type": "number" },
                    "stock": { "type": "integer" }
                  },
                  "required": ["_id", "name", "brand", "category_id", "price", "stock"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Sneaker not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a sneaker by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the sneaker to delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Sneaker deleted"
          },
          "404": {
            "description": "Sneaker not found"
          }
        }
      }
    }
  }
}